requirements.txt
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==1.2.0
fonttools==4.47.2
idna==3.6
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pybit==5.6.2
pycryptodome==3.20.0
Pygments==2.16.1
pyparsing==3.1.1
pyTelegramBotAPI==4.15.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
schedule==1.2.1
six==1.16.0
stack-data==0.6.2
tabulate==0.9.0
telebot==0.0.5
tornado==6.3.3
traitlets==5.10.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.6
websocket-client==1.7.0
websockets==12.0
